Удаление временных файлов

+++ src/ArchiveDoc/cntDocuments.cs
@@ -545,7 +545,12 @@ namespace ArchiveDoc
                 string fileName = (string)task.Result.Rows[0]["FileName"];
                 if (!Directory.Exists(Application.StartupPath + @"\tmp\"))
                     Directory.CreateDirectory(Application.StartupPath + @"\tmp\");
-                File.WriteAllBytes(Application.StartupPath + @"\tmp\" + fileName, file);
+                //File.WriteAllBytes(Application.StartupPath + @"\tmp\" + fileName, file);
+                //Process.Start(Application.StartupPath + @"\tmp\" + fileName);
+                if (!File.Exists(Application.StartupPath + @"\tmp\" + fileName))
+                {
+                    File.WriteAllBytes(Application.StartupPath + @"\tmp\" + fileName, file);                    
+                }
                 Process.Start(Application.StartupPath + @"\tmp\" + fileName);
             }
         }
